+++ src/app/aplication-state.ts
@@ -1,5 +1,5 @@
-import { IProductsInitialState } from './products/store/product.reducer';
-import { IUsersInitialState } from './users/store/user.reducer';
+import { IProductsInitialState } from './products/ngrx/product.reducer';
+import { IUsersInitialState } from './users/ngrx/user.reducer';
 
 export interface IApplicationState {
   userState: IUsersInitialState;
