+++ src/app/products/ngrx/product.reducer.ts
