+++ src/app/products/products-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ProductsComponent } from './products.component';
+import { ProductsListComponent } from './product-list/products-list.component';
 
 const routes: Routes = [
   {
     path: 'products',
-    component: ProductsComponent,
+    component: ProductsListComponent,
   },
   // { path: 'users/:id/edit', component: UserEditComponent },
   // { path: 'users/new', component: UserEditComponent },
