+++ src/app/users/users.module.ts
@@ -1,5 +1,8 @@
 import { NgModule } from '@angular/core';
-import { UserDeleteConfirm, UsersComponent } from './users.component';
+import {
+  UserDeleteConfirm,
+  UsersListComponent,
+} from './users-list/users-list.component';
 import { UserEditComponent } from './user-edit/user-edit.component';
 import { RouterModule } from '@angular/router';
 import { UsersRoutingModule } from './users-routing.module';
@@ -7,12 +10,12 @@ import { SharedModule } from '../shared/shared.module';
 import { MaterialModule } from '../shared/material.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { StoreModule } from '@ngrx/store';
-import * as fromUser from './store/user.reducer';
+import * as fromUser from './ngrx/user.reducer';
 import { EffectsModule } from '@ngrx/effects';
-import { UserEffects } from './store/user.effects';
+import { UserEffects } from './ngrx/user.effects';
 
 @NgModule({
-  declarations: [UsersComponent, UserEditComponent, UserDeleteConfirm],
+  declarations: [UsersListComponent, UserEditComponent, UserDeleteConfirm],
   imports: [
     RouterModule,
     UsersRoutingModule,
